travidox_backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
firebase-admin==6.2.0
paramiko==3.3.1
pyzmq==25.1.1
python-dotenv==1.0.0
pydantic==2.4.2
typing-extensions==4.8.0
httpx==0.25.0
pytest==7.4.3
pytest-asyncio==0.21.1
requests==2.31.0
cryptography==41.0.5
bcrypt==4.0.1